\docType{package}
\name{h0-package}
\alias{h0-package}
\title{A Bayesian Meta-Analysis for Estimating the Hubble Constant via Time Delay Cosmography}
\description{
The R package \pkg{h0} provides a toolbox to conduct a Bayesian meta-analysis for estimating the current expansion rate of the Universe, called the Hubble constant H0, via time delay cosmography. The input data are (i) estimates of Fermat potential differences and their one-sigma uncertainties (standard errors), and (ii) estimates of time delays and their one-sigma uncertainties. Given these inputs, the meta-analysis draws a posterior sample of the model parameters including the Hubble constant.
}
\details{
\tabular{ll}{
Package: \tab h0\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
Date: \tab 2023-8-24\cr
License: \tab GPL-2\cr
Main functions: \tab \code{\link{h0}}\cr
}
}
\author{
Hyungsuk Tak
}
\references{
    T. Treu and P. Marshall (2016) <doi:10.1007/s00159-016-0096-8> "Time Delay Cosmography".
}
